feat(api): enhance ChatModelFactory for OpenAI compatibility and improve prompt handling in tests

+++ pulsar-tests/src/test/kotlin/ai/platon/pulsar/test2/browser/PulsarWebDriverTests.kt
@@ -110,7 +110,7 @@ class PulsarWebDriverTests : WebDriverTestBase() {
         assertEquals(1920, metadata.viewPortWidth)
         assertEquals(1080, metadata.viewPortHeight)
         assertTrue { metadata.scrollTop > 0 }
-        assertTrue { metadata.scrollLeft == 0 }
+        assertTrue { metadata.scrollLeft.toInt() == 0 }
         assertTrue { metadata.clientWidth > 0 } // 1683 on my laptop
         assertTrue { metadata.clientHeight > 0 } // 986 on my laptop
     }
